pyproject.toml
[project]
name = "compliance-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.136.0",
    "google-generativeai>=0.8.6",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.2",
    "langgraph>=1.1.8",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "streamlit>=1.56.0",
    "uvicorn>=0.44.0",
    "yt-dlp>=2026.3.17",
]
